My solution Thomas Bourany
I tried to do the homework and after a (very) long time trying to solve the errors, I looked at the code in solution. It seems that I got everything wrong (from the beginning to the end). So, here are the stuff I could do, I know it does not work, but I tried (I need to devote some time for the next homework!).